requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==24.2.0
audioread==3.0.1
azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.22.0
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bidict==0.22.1
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
blinker==1.6.2
blis==0.7.9
boto3==1.35.52
botocore==1.35.52
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.16.0
chainlit==0.7.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.71
collection==0.1.6
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cryptography==43.0.1
ctransformers==0.2.27
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
datasets==2.15.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.7
diskcache==5.6.3
docopt==0.6.2
docx==0.2.4
docx2txt==0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl#sha256=8902305d2ced83d98a8e88efd93ea8970a70dda9bb24b0024a8b798d1cc913d5
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.99.1
fastapi-socketio==0.0.10
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
flatbuffers==23.5.26
fonttools==4.41.0
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gensim==4.3.1
gingerit==0.9.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.22.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-texttospeech==2.20.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
gprofiler-official==1.0.0
gradio==3.45.2
gradio_client==0.5.3
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.19.4
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
indic-transliteration==2.3.44
ipykernel==6.24.0
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
langchain==0.0.304
langcodes==3.3.0
langsmith==0.0.41
Lazify==0.4.0
lazy_loader==0.3
Levenshtein==0.21.1
lexical-diversity==0.1.1
libclang==16.0.0
librosa==0.10.1
lightning-utilities==0.9.0
limits==3.13.0
llama_cpp_python==0.2.7
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
murmurhash==1.0.9
mygene==3.2.2
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numba==0.58.1
numexpr==2.8.7
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.7
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.1
pooch==1.7.0
preshed==3.0.8
prisma==0.10.0
prompt-toolkit==3.0.39
proto-plus==1.25.0
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
pydub==0.25.1
pyenchant==3.2.2
Pygments==2.15.1
pygrams==0.0.14
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyre-extensions==0.0.29
PySocks==1.7.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-engineio==4.7.1
python-graphql-client==0.4.3
python-Levenshtein==0.21.1
python-multipart==0.0.6
python-socketio==5.9.0
pytorch-lightning==2.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==3.1.1
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.2
roman==4.1
rpds-py==0.8.10
rsa==4.9
s3transfer==0.10.3
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.25.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
spacy==3.5.4
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.5
SQLAlchemy==2.0.21
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.24.1
sumy==0.11.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
textblob==0.17.1
thinc==8.1.10
thinc-apple-ops==0.1.3
threadpoolctl==3.1.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.0.0
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.35.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
uptrace==1.20.0
urllib3==1.26.16
uvicorn==0.23.2
validators==0.20.0
wasabi==1.1.2
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.3.6
word2number==1.1
wrapt==1.15.0
wsproto==1.2.0
xformers==0.0.20
xgboost==1.7.6
xxhash==3.4.1
yarl==1.9.2
yfinance==0.2.24
zipp==3.16.1
